Make the following changes:
,

Find this element:

  <Primitive.div role="dialog" id="radix-«r0»" aria-describedby="radix-«r2»" aria-labelledby="radix-«r1»" data-state="open" className="fixed z-50 gap-4 shadow-lg transition ease-in-out data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:duration-300 data-[state=open]:duration-500 inset-y-0 left-0 h-full data-[state=closed]:slide-out-to-left data-[state=open]:slide-in-from-left sm:max-w-sm w-[280px] sm:w-[320px] p-0 bg-[hsl(270,60%,75%)] border-r border-[hsl(270,60%,65%)]" tabIndex=-1 style={"pointerEvents":"auto"}>

with the text content:

    MenuMendspace HomeRecent SubmissionsAll EntriesClose

Relevant files:

    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_c7142e05._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_c7142e05._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_fa006f38._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/src_75ad5782._.js
    - https://6000-firebase-studio-1747321423346.cluster-ancjwrkgr5dvux4qug5rbzyc2y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

add this below the menu  Terms and Conditions:- The parties involved do not intend on attack or blame or to defend , they just intend on vocalising their feelings. They intend on feeling heard , and their needs being met . Anybody’s feeling is not to be taken as an attack , as it is not, it's merely an unmet need.  Refrain from trying to fix your language or feeling, anything you feel is something you write .  Use of any language, to fill this is allowed . Both parties are required to fill this up after every argument/disagreement because they love each other and are willing to make this work. Rules :- In the table given below , where you have three columns , in one side you're asked to fill up,(i) what the other person said ,(ii) what you felt ,(iii) in the defense of the other  person .  The first two columns are to be filled by one and the third column is to be filled by the other participant .  You can customise your table with any colour/font of your liking . After this is filled completely , a discussion is to be followed where both parties are given space and time to i) to feel justified or ii)talk about things that bother them .  Beside your signature , do drop an emoji of your liking .  This is to be filled at the soonest possible time convenient after a conflict of interest